+++ packages/sdk-ts/src/core/accounts/PrivateKey.spec.ts
@@ -1,18 +1,17 @@
-import { verifyMessage, Wallet } from 'ethers'
-import { toUtf8 } from '../../utils/utf8.js'
+import { verifyMessage, Wallet, hashMessage } from 'ethers'
 import { PrivateKey } from './PrivateKey.js'
 import { generateArbitrarySignDoc } from '../tx/index.js'
 
 const pk = process.env.TEST_PRIVATE_KEY as string
-const seedPhase = process.env.TEST_SEED_PHASE as string
+const seedPhrase = process.env.TEST_SEED_PHRASE as string
 
-if (!pk || !seedPhase) {
-  throw new Error('TEST_PRIVATE_KEY or TEST_SEED_PHASE is not set')
+if (!pk || !seedPhrase) {
+  throw new Error('TEST_PRIVATE_KEY or TEST_SEED_PHRASE is not set')
 }
 
 describe('PrivateKey', () => {
   it('returns the correct address derived from a mnemonic', () => {
-    const privateKey = PrivateKey.fromMnemonic(seedPhase)
+    const privateKey = PrivateKey.fromMnemonic(seedPhrase)
 
     expect(privateKey.toBech32()).toEqual(
       'inj1u3f7yuwl8mkd88ryzdhxt6kt90a60rgl0nmxyq',
@@ -121,26 +120,25 @@ describe('PrivateKey', () => {
 
   it('returns true when checking a pk signature against the signer public key', async () => {
     const message = 'this is a test message'
+    const messageHash = hashMessage(message)
 
     const wallet = new Wallet(pk)
-    const ethersSignature = await wallet.signMessage(message)
-
     const privateKey = PrivateKey.fromHex(pk)
-    const publicKey = privateKey.toHex()
-
+    
+    const ethersSignature = await wallet.signMessage(message)
     const ethersVerifiedSigner = verifyMessage(message, ethersSignature)
-    const ethersSignatureVerifiedCorrectly = ethersVerifiedSigner === publicKey
+    const ethersSignatureVerifiedCorrectly = ethersVerifiedSigner.toLowerCase() === wallet.address.toLowerCase()
     expect(ethersSignatureVerifiedCorrectly).toBe(true)
 
     const privKeySignatureArray = privateKey.signHashed(
-      Buffer.from(toUtf8(message), 'utf-8'),
+      Buffer.from(messageHash.slice(2), 'hex')
     )
     const privKeySignature = `0x${Buffer.from(privKeySignatureArray).toString(
       'hex',
     )}`
     const privKeyVerifiedSigner = verifyMessage(message, privKeySignature)
     const privKeySignatureVerifiedCorrectly =
-      privKeyVerifiedSigner === publicKey
+      privKeyVerifiedSigner.toLowerCase() === wallet.address.toLowerCase()
     expect(privKeySignatureVerifiedCorrectly).toBe(true)
   })