+++ packages/sdk-ts/src/core/modules/wasm/exec-args/ExecArgCreateSpotGridStrategy.ts
@@ -27,6 +27,7 @@ export declare namespace ExecArgCreateSpotGridStrategy {
       lowerTrailing: string
       lpMode?: boolean
     }
+    feeRecipient?: string
   }
 
   export interface Data {
@@ -44,12 +45,14 @@ export declare namespace ExecArgCreateSpotGridStrategy {
     }
     exit_type?: ExitType
     strategy_type?: StrategyType | TrailingArithmetic | TrailingArithmeticLP
+    fee_recipient?: string
   }
 }
 
 /**
  * @category Contract Exec Arguments
  */
+/** @deprecated */
 export default class ExecArgCreateSpotGridStrategy extends ExecArgBase<
   ExecArgCreateSpotGridStrategy.Params,
   ExecArgCreateSpotGridStrategy.Data
@@ -96,6 +99,9 @@ export default class ExecArgCreateSpotGridStrategy extends ExecArgBase<
           }
         : undefined,
       strategy_type: strategyType,
+      ...(params.feeRecipient && {
+        fee_recipient: params.feeRecipient,
+      }),
     }
   }