+++ packages/sdk-ts/src/service/index.ts
@@ -1,2 +1,3 @@
 export * from './TokenPrice.js'
 export * from './TokenFactoryStatic.js'
+export * from './TokenStaticFactory.js'